requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
customtkinter==5.2.0
darkdetect==0.8.0
decorator==5.1.1
future==0.18.3
geocoder==1.38.1
idna==3.4
Pillow==10.0.0
pyperclip==1.8.2
pywin32==306
ratelim==0.1.6
requests==2.31.0
six==1.16.0
tk==0.1.0
tkintermapview==1.29
typing_extensions==4.7.1
urllib3==2.0.4
